Fix disjoint sets

+++ bolt_core/src/bolt_core/SparseStorage.h
@@ -128,7 +128,7 @@ public:
     }
 
     std::pair<unsigned int, unsigned int> endpoints_;
-    double weight_;
+    float weight_;
   };
 
   /** \brief Constructor */
